chore(web): refactor react code
