+++ web/src/containers/Studio/internal/helper.ts
@@ -1,0 +1,14 @@
+export const getLanguageFromExtension = (extension: string) => {
+  if (extension === 'js') return 'javascript'
+
+  if (extension === 'ts') return 'typescript'
+
+  if (extension === 'md' || extension === 'mdx') return 'markdown'
+
+  return extension
+}
+
+export const getSelection = (editor: any) => {
+  const selection = editor?.getModel().getValueInRange(editor?.getSelection())
+  return selection ?? ''
+}
