+++ web/src/utils/hooks/index.ts
@@ -1,2 +1,4 @@
+export * from './useModal'
 export * from './usePrompt'
 export * from './useStateWithCallback'
+export * from './useSnackbar'
